Move SlotMap into collections package
