+++ foresight/src/main/scala/foresight/eqsat/extraction/CostAnalysis.scala
@@ -1,6 +1,7 @@
 package foresight.eqsat.extraction
 
 import foresight.eqsat._
+import foresight.eqsat.collections.SlotMap
 import foresight.eqsat.metadata.{Analysis, EGraphWithMetadata}
 
 /**
