+++ foresight/src/main/scala/foresight/eqsat/hashCons/MutableSlottedUnionFind.scala
@@ -1,7 +1,7 @@
 package foresight.eqsat.hashCons
 
-import foresight.eqsat.collections.SlotSet
-import foresight.eqsat.{EClassCall, EClassRef, Slot, SlotMap}
+import foresight.eqsat.collections.{SlotMap, SlotSet}
+import foresight.eqsat.{EClassCall, EClassRef, Slot}
 
 private final class MutableSlottedUnionFind(var parents: Map[EClassRef, EClassCall]) {
   def toImmutable: SlottedUnionFind = SlottedUnionFind(parents)
